requirements.txt
torchio==0.20.19
numpy==1.26.4
nibabel==5.3.2
einops==0.8.1
tensorboard==2.20.0
scikit-learn==1.7.1
matplotlib==3.10.5
pandas==2.3.1
imageio==2.37.0
monai==1.4.0